Extract st.container, columns, container out of delta_generator.py
